Guard process faster stop
The guard process sleeps for the guard cycle interval in each
iteration before checking the stop event. It is more effective
to use the event's wait() method rather than time.sleep() so
the process wakes immediately when the event is set.

The commit also fixes the case where the guard process sleeps for
an extra cycle when the cycle counter happens to have been reset to
zero on the loop iterations when the scheduler is called.

This fix helps in our environment where we are frequently stopping
and starting the cluster and where we have increased the guard
cycle setting to several seconds.